% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_helpers.R
\name{get_checkers}
\alias{get_checkers}
\title{get_checkers}
\usage{
get_checkers(raw_dat)
}
\arguments{
\item{raw_dat}{Initial raw data}
}
\value{

}
\description{
get_checkers
}
\examples{
\dontrun{
}
}
