% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_helpers.R
\name{parse_education_data}
\alias{parse_education_data}
\title{parse_education_data}
\usage{
parse_education_data(filename)
}
\arguments{
\item{filename}{filename}
}
\value{
ggplot2 object
}
\description{
Ggplot2 Age viz
}
\examples{
\dontrun{
}
}
