% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/article_scrips.R
\name{plot_age_viz}
\alias{plot_age_viz}
\title{plot_age_viz}
\usage{
plot_age_viz(variables)
}
\arguments{
\item{variables}{a character string.}
}
\value{
ggplot2 object
}
\description{
Ggplot2 Age viz
}
\examples{
\dontrun{
}
}
