% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/article_scrips.R
\name{plot_educ_viz}
\alias{plot_educ_viz}
\title{plot_educ_viz}
\usage{
plot_educ_viz()
}
\arguments{
\item{variables}{a character string.}
}
\value{
ggplot2 object
}
\description{
Ggplot2 Education viz
}
\examples{
\dontrun{
}
}
